move: global ui as shared ui
