+++ apps/web/src/components/ui/ExerciseData/ExerciseDataList.tsx
@@ -1,7 +1,7 @@
 import type { Schedule } from 'fitness-struct';
 import { Suspense } from 'react';
 import { useGetExerciseListByScheduleId } from '@hooks/apollo/Exercise';
-import MenuableAccordion from '@ui/CustomComponent/MenuableAccordion';
+import MenuableAccordion from '@shared/ui/MenuableAccordion';
 import SetListEditor from '../Sets/SetListEditor';
 
 export interface ExerciseDataListProps {
