+++ apps/web/src/components/ui/Preset/PresetDisplay.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { useExercisePresetWithListFragment } from '@hooks/apollo/ExercisePreset';
 import usePresetMenu from '@hooks/usePresetMenu';
-import MenuableAccordion from '@ui/CustomComponent/MenuableAccordion';
+import MenuableAccordion from '@shared/ui/MenuableAccordion';
 import SimpleFitnessList from '@ui/Fitness/SimpleFitnessList';
 
 interface PresetDisplayProps {
