+++ apps/web/src/pages/Main.tsx
@@ -5,8 +5,8 @@ import { useNavigate } from 'react-router-dom'
 import { useGetScheduleByDate } from '@hooks/apollo/Schedule'
 import { useScheduleActions } from '@hooks/useScheduleMenu'
 import usePageTracker from '@shared/hooks/usePageTracker'
+import MenuableAccordion from '@shared/ui/MenuableAccordion'
 import StateRender from '@shared/ui/StateRender'
-import MenuableAccordion from '@ui/CustomComponent/MenuableAccordion'
 import ScheduleDisplay from '@ui/Schedule/ScheduleDisplay'
 import { useBottomNavi } from '@widgets/bottomNavi'
 import { DateService } from '@widgets/calender'
